requirements.txt
pipecat-ai
torch
torchaudio
soundfile
aiohttp
aiofiles
websockets
